Add TestValidation class and move GlobalsOpimizedAway test
