+++ dawn/test/unit-test/dawn/Optimizer/Passes/CMakeLists.txt
@@ -23,6 +23,7 @@ add_executable(${executable}
   TestFieldAccessIntervals.cpp
   TestTemporaryToFunction.cpp
   TestIntervalPartitioner.cpp
+  TestValidation.cpp
 )
 target_link_libraries(${executable} PRIVATE DawnOptimizer DawnCompiler DawnUnittest gtest)
 target_add_dawn_standard_props(${executable})
