+++ dawn/test/unit-test/dawn/Optimizer/Passes/globals_opt_away.sir
